+++ client/src/App.tsx
@@ -22,6 +22,7 @@ import { useWebVitals } from "@/lib/use-web-vitals";
 import { CircularProgress } from "@/components/ui/circular-progress";
 import { LoadingProvider } from "@/lib/loading-context";
 import { GlobalQuickActions } from "@/components/GlobalQuickActions";
+import { WishlistProvider } from "@/lib/wishlist";
 
 const Landing = lazy(() => import("@/pages/Landing"));
 const Login = lazy(() => import("@/pages/Login"));
@@ -400,20 +401,22 @@ function App() {
 
         <ThemeProvider defaultTheme="light" switchable>
           <LoadingProvider>
-            <OfflineProvider>
-              <OfflineBanner />
-              <TooltipProvider>
-                <Toaster />
-                <Router />
-                <ScrollProgress />
-                <CommandPalette />
-                <GlobalQuickActions />
-                <FloatingSupportWidget />
-                <AliasAIAssistant />
-                <InstallPrompt />
-                <SWUpdateToast />
-              </TooltipProvider>
-            </OfflineProvider>
+            <WishlistProvider>
+              <OfflineProvider>
+                <OfflineBanner />
+                <TooltipProvider>
+                  <Toaster />
+                  <Router />
+                  <ScrollProgress />
+                  <CommandPalette />
+                  <GlobalQuickActions />
+                  <FloatingSupportWidget />
+                  <AliasAIAssistant />
+                  <InstallPrompt />
+                  <SWUpdateToast />
+                </TooltipProvider>
+              </OfflineProvider>
+            </WishlistProvider>
           </LoadingProvider>
         </ThemeProvider>
       </ErrorBoundary>